fix: mover un controlador de un helper a otro.
