+++ controllers/entrada.go
@@ -55,7 +55,6 @@ func (c *EntradaController) Post() {
 
 		logs.Info("Lo que llega", v)
 		if respuesta, err := entradaHelper.AddEntrada(v); err == nil && respuesta != nil {
-			//		if respuesta, err := entradaHelper.MvtoContableEntrada(v, "Comprobante entrada"); err == nil && respuesta != nil {
 			c.Ctx.Output.SetStatus(201)
 			c.Data["json"] = respuesta
 		} else {
