+++ models/consecutivos_crud.go
@@ -1,0 +1,10 @@
+package models
+
+type Consecutivo struct {
+	Id          int
+	ContextoId  int
+	Year        int
+	Consecutivo int
+	Descripcion string
+	Activo      bool
+}
